% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_freq_boundaries.R
\name{geno_freq_boundaries}
\alias{geno_freq_boundaries}
\title{function to return a tibble with the min/max values of genotype freqs possible}
\usage{
geno_freq_boundaries()
}
\value{
a tibble
}
\description{
These mins and maxes occur because the genotypes are used to estimate
the allele frequencies.  The output of this function is useful for
putting boundaries on plots.
}
\examples{
gfb <- geno_freq_boundaries()
}
\keyword{internal}
