% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_freq_count_single.R
\name{geno_freq_calc_single}
\alias{geno_freq_calc_single}
\title{from on 012 file compute expected (assuming HW equilbrium) and observed genotype counts}
\usage{
geno_freq_calc_single(g012)
}
\arguments{
\item{g012}{an 012 matrix, with indivs in rows and loci in columns.
Missing data can be -1 or NA. The matrix must
have colnames which are the locus names.}
}
\value{
See information for return value of \code{\link{exp_and_obs_geno_freqs}}.
}
\description{
This is an internal function.
}
\examples{
# get an 012 matrix from the lobster data
tmp <- t(vcfR::extract.gt(lobster_buz_2000, element = "GT"))
locnames <- colnames(tmp)
g <- make_it_012(tmp)
colnames(g) <- locnames # put these back on since make_it_012 removes them
gf <- geno_freq_calc_single(g)
}
\keyword{internal}
