% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsM}
\alias{gibbsM}
\title{simulate a new miscall rate for each read depth category given X and Y}
\usage{
gibbsM(M, num_cats, X, Y, R, pri)
}
\description{
This just writes new values into M as if it were an output variable
}
\keyword{internal}
