% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsP}
\alias{gibbsP}
\title{simulate new reference allele frequencies from their beta full conditional}
\usage{
gibbsP(p, X, pri)
}
\description{
This just writes new values into P as if it were an output variable
}
\keyword{internal}
