% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsX}
\alias{gibbsX}
\title{compute full conditional for each X given Y, p, R, and m, and then sample from it}
\usage{
gibbsX(X, Y, R, p, M)
}
\description{
This just writes new values into X as if it were an output variable.
Internal function used within \code{\link{estimate_m_rd}}.
}
\keyword{internal}
