% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lobster_buz_2000}
\alias{lobster_buz_2000}
\title{RAD-seq data from 36 lobsters at 2000 SNPs}
\format{A vcfR object in which the original data set has
been reduced to just the 36 lobsters from the BUZ population
and a randomly sampled 2000 SNPs from the 10,156 originally
available.

This is the sort of object obtained after calling \code{vcfR::read.vcfR()}
on a VCF file.}
\source{
\url{https://datadryad.org//resource/doi:10.5061/dryad.q771r.3}
}
\usage{
lobster_buz_2000
}
\description{
A data set from the study by Benestan et al. (2016).
}
\keyword{datasets}
