% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lobster_buz_2000_as_012_matrix}
\alias{lobster_buz_2000_as_012_matrix}
\title{An 012 matrix from  (RAD) sequence data from 36 lobsters at 2000 single nucleotide polymorphisms (SNPs)}
\format{
This is an integer matrix with positions in rows and samples in columns.
It is an 012 matrix that corresponds to lobster_buz_2000.  It is useful as
an example of the necessary format for the d012 argument to \code{\link{exp_and_obs_geno_freqs}}.
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.q771r}
}
\usage{
lobster_buz_2000_as_012_matrix
}
\description{
A data set from the study by Benestan et al. (2016).
}
\keyword{datasets}
