% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_vcf_for_est_m_rd.R
\name{prep_vcf_for_est_m_rd}
\alias{prep_vcf_for_est_m_rd}
\title{convert a VCF into an 012,-1 matrix and read_depth bin matrix for estimation}
\usage{
prep_vcf_for_est_m_rd(v, DF, minBin)
}
\arguments{
\item{v}{a vcfR object into which a VCF file has been read}

\item{DF}{Field to use for obtaining total read depth.  Choices are DP and AD, but only DP is
implemented at this point. And you have to make sure that DP is a field that exists...}

\item{minBin}{minimum number of observations for each read depth bin}
}
\value{
This sends back a list with \code{mat012}: the 012,-1 matrix of genotypes, and
\code{dp_bins_list}: the list returned by bin_depths().
}
\description{
convert a VCF into an 012,-1 matrix and read_depth bin matrix for estimation
}
\keyword{internal}
