% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_whoa.R
\name{read_whoa}
\alias{read_whoa}
\title{read vcf file into whoa input file}
\usage{
read_whoa(data, indivs = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{(character string) The VCF file. SNPs are biallelic or haplotypes.}

\item{indivs}{(optional, character) Use this argument to select
a subsample of id in the vcf file, e.g. a particular group or population of samples.
The file is a tab-separated file with 1 column, no header, containing the individuals to
include in the analysis.
By default, this is NULL, in which case everyone from the file is included.
Default: \code{indivs = NULL}.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
whoa will append \code{.gds} to the filename.
If filename chosen is already present in the
working directory, the default \code{whoa_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) Default: \code{parallel::detectCores() - 1}.
The number of processor used during execution.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function (see details).}
}
\description{
read vcf file into whoa \href{https://github.com/eriqande/whoa}{whoa}
package.
}
\details{
\strong{Advance mode, using \emph{dots-dots-dots ...}}
\enumerate{
\item \code{sample.snps}: (optional, integer) Select a random number of SNPs
to speed up analysis, e.g. using \code{sample.snps = 1000}.
Default: \code{sample.snps = NULL}, will use all SNPs.
\item \code{random.seed} (optional, integer) For reproducibility, set an integer
that will be used inside function that requires randomness. With default,
a random number is generated.
Default: \code{random.seed = NULL}.
\item \code{path.folder}: to write ouput in a specific path
(used internally in radiator). Default: \code{path.folder = getwd()}.
If the supplied directory doesn't exist, it's created.
}
}
\examples{
\dontrun{
# with built-in defaults:
 data <- whoa::read_whoa(data = "populations.snps.vcf")

# Using more arguments:
data <- whoa::read_whoa(
data = "populations.snps.vcf",
indivs = "select_ind.tsv",
parallel.core = 5,
filename = "salamander",
path.folder = "salamander/het_analysis")
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
