% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_whoa.R
\name{run_whoa}
\alias{run_whoa}
\title{Wrapper function that runs whoa}
\usage{
run_whoa(data, indivs = NULL, strata = NULL,
  prop_indv_required = 0.5, prop_loci_required = 0.5, alpha = 0.2,
  max_plot_loci = 500, minBin, init_m = 0.1, num_sweeps = 500,
  burn_in = 100, parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{(character string) The VCF file. SNPs are biallelic or haplotypes.}

\item{indivs}{(optional, character) Use this argument to select
a subsample of id in the vcf file, e.g. a particular group or population of samples.
2 options: use a character string with desired vcf sample id or a
tab-separated file with 1 column, no header, containing the individuals to
include in the analysis.
By default, this is NULL, in which case everyone from the file is included.
Default: \code{indivs = NULL}.}

\item{strata}{(optional) A tab-separated file with 2 columns: \code{INDIVIDUALS} and \code{STRATA}.
Using this argument will run the function serially on all the strata and overall.
Default: \code{strata = NULL}.}

\item{prop_indv_required}{loci will be dropped if a proportion of
individuals less than prop_indv_required have non-missing data at that locus.
Default is 0.5}

\item{prop_loci_required}{individual will be dropped if their proportion of
non-missing loci is less than prop_loci_required. Default is 0.5.}

\item{alpha}{the transparency (alpha) parameter to apply to the points
in the scatterplot. Default is 0.2.}

\item{max_plot_loci}{By default this plots only 500 loci, sampled
randomly, to keep ggplot taking forever to plot, for example, 100K
points.  If you want to plot all the points, set this to a number
larger than the number of SNPs in the data set.}

\item{minBin}{minimum number of observations for each read depth bin.  If you have 10K markers
and 50 individuals then you have about 500,000 genotypes to play with.  Requiring bins with at
least 5,000 genotypes in them will give you less than 100 bins.  You can play around with this
number to get the right number of bins.  The algorithm breaks the read depths up into bins that
have at least minBin genotypes in them.}

\item{init_m}{the initial value of the heterozygote miscall rate assumed for each read depth bin.
By default this is 0.1.}

\item{num_sweeps}{the number of sweeps of the MCMC algorithm to do.  Default is 500, which is a
little on the short side.  Run multiple times and make sure the values obtained are similar across
runs to assess convergence.}

\item{burn_in}{how many sweeps from the beginning of the chain to discard when computing the
posterior means and quantiles.  Default is 100.  Note that full traces of the visited m values
for every read depth bin are returned as well, so that the behavior of the chain in those early
steps can be investigated.}

\item{parallel.core}{(optional) Default: \code{parallel::detectCores() - 1}.
The number of processor used during execution.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function (see details).}
}
\value{
When strata argument is used, a list with:
\enumerate{
\item \code{gds}: the gds object
\item \code{het.miscall.rate.analysis}: a list with \code{gfreqs},
\code{plot.freqs}, \code{binned}, \code{plot.post} for each strata.
\item \code{random.seed}: the random seed used
\item \code{overall.rate}: the overall heterozygote miscall rate (across read depth and strata)
}
}
\description{
Wrapper function that runs \href{https://github.com/eriqande/whoa}{whoa} on a vcf.
When the strata argument is used, the heterozygote miscall rate is calculated for each strata.
}
\details{
\strong{Advance mode, using \emph{dots-dots-dots ...}}
\enumerate{
\item \code{sample.snps}: (optional, integer) Select a random number of SNPs
to speed up analysis, e.g. using \code{sample.snps = 1000}.
Default: \code{sample.snps = NULL}, will use all SNPs.
\item \code{random.seed} (optional, integer) For reproducibility, set an integer
that will be used inside function that requires randomness. With default,
a random number is generated.
Default: \code{random.seed = NULL}.
}
}
\examples{
\dontrun{
require(httr)
require(ggpubr)
library(whoa)
# get the vcf and strata files:
writeBin(httr::content(httr::GET("https://datadryad.org/bitstream/handle/
10255/dryad.108679/10156-586.recode.vcf?sequence=1"), "raw"),"lobster_data_2015.vcf")
writeBin(httr::content(httr::GET("https://datadryad.org/bitstream/handle/
10255/dryad.108679/README.txt?sequence=2"), "raw"),"lobster_strata_2015.tsv")
# work on the strata:
readr::read_tsv(file = "lobster_strata_2015.tsv",
               col_names = c("INDIVIDUALS", "STRATA"),
               col_types = "cc") \%>\%
 dplyr::mutate(
   STRATA = stringi::stri_sub(str = INDIVIDUALS, from = 1, to = 3),
   STRATA = stringi::stri_replace_all_fixed(
     str = STRATA,
     pattern = c("CAP", "EDN", "SID"),
     replacement = c("MAG", "MAG", "DIN"),
     vectorize_all = FALSE)
 ) \%>\%
 dplyr::filter(STRATA != "TON") \%>\%
 dplyr::arrange(STRATA, INDIVIDUALS) \%>\%
 readr::write_tsv(x = ., path = "lobster_strata_2015.tsv")

 # run whoa on all strata
 het.miscall.rate.pop <- run_whoa(
     data = "lobster_data_2015.vcf", strata = "lobster_strata_2015.tsv",
     sample.snps = 2000,
     max_plot_loci = 2000,
     minBin = 2000)


# To view the combined ggplot of the posterior estimates for heterozyote miscall rates
fig <- purrr::map(.x = het.miscall.rate.pop$het.miscall.rate.analysis, .f = 4) \%>\%
    ggpubr::ggarrange(plotlist = ., legend = "right", common.legend = TRUE,
        labels = names(.), label.x = 0.5, label.y = 0.9)
fig
}
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov} and
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
