% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_m_ests.R
\name{tidy_m_ests}
\alias{tidy_m_ests}
\title{tidy up the estimate_m_rd output into something you can plot}
\usage{
tidy_m_ests(E, S, TB, burn = 50)
}
\arguments{
\item{E}{the list returned by the estimation function}

\item{S}{the bin stats}

\item{TB}{the tidy bins}

\item{burn}{how much for burn in?}
}
\description{
tidy up the estimate_m_rd output into something you can plot
}
\keyword{internal}
