% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoa.R
\docType{package}
\name{whoa}
\alias{whoa}
\alias{whoa-package}
\title{whoa: Evaluation of genotyping error in genotype by sequencing data}
\description{
The name is found in the capitals here:  \code{W}here's my \code{H}eterozygotes?! \code{O}bservations
of genotyping \code{A}ccuracy.
}
\section{the \code{whoa} main high-level functions}{


The main function in the package whoa is \code{\link{infer_m}}.
This function infers the heterozygote miscall rate (the rate at
which true heterozygotes have been miscalled as homozygotes in
genotype-by-sequencing data) for calls made upon genotypes falling
within different read depth categories.

The output from  \code{\link{infer_m}} is easily plotted by passing the
m_posteriors component of the output list from \code{infer_m} into
\code{\link{posteriors_plot}}.
}

\section{example data}{


The package comes with a small data set, \code{\link{lobster_buz_2000}}, which
was read in from a variant call format (VCF) file and is now stored in the package as a `vcfR' object.
}

