% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_whoa.R
\name{read_whoa}
\alias{read_whoa}
\title{read vcf file for whoa}
\usage{
read_whoa(data, indivs = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, ...)
}
\arguments{
\item{data}{(character string) The VCF file. SNPs are biallelic or haplotypes.}

\item{indivs}{(optional, character) Use this argument to select
a subsample of id in the vcf file, e.g. a particular group or population of samples.
2 options: use a character string with desired vcf sample id or a
tab-separated file with 1 column, no header, containing the individuals to
include in the analysis.
By default, this is NULL, in which case everyone from the file is included.
Default: \code{indivs = NULL}.}

\item{filename}{(optional) The file name of the Genomic Data Structure (GDS) file.
whoa will append \code{.gds} to the filename.
If filename chosen is already present in the
working directory, the default \code{whoa_datetime.gds} is chosen.
Default: \code{filename = NULL}.}

\item{parallel.core}{(optional) Default: \code{parallel::detectCores() - 1}.
The number of processor used during execution.}

\item{...}{(optional) Advance mode that allows to pass further arguments
for fine-tuning the function (see details).}
}
\value{
A gds object.
}
\description{
read and prepare the vcf file for whoa \href{https://github.com/eriqande/whoa}{whoa}
package.
}
\details{
\strong{Advance mode, using \emph{dots-dots-dots ...}}
\enumerate{
\item \code{sample.snps}: (optional, integer) Select a random number of SNPs
to speed up analysis, e.g. using \code{sample.snps = 1000}.
Default: \code{sample.snps = NULL}, will use all SNPs.
\item \code{random.seed} (optional, integer) For reproducibility, set an integer
that will be used inside function that requires randomness. With default,
a random number is generated.
Default: \code{random.seed = NULL}.
}
}
\examples{
\dontrun{
require(httr)
require(SeqArray)
library(whoa)

# Get the lobster vcf file:
writeBin(httr::content(httr::GET("https://datadryad.org/bitstream/handle/
10255/dryad.108679/10156-586.recode.vcf?sequence=1"), "raw"),"lobster_data_2015.vcf")

# Get the strata file containing individuals and groupings
writeBin(httr::content(httr::GET("https://datadryad.org/bitstream/handle/
10255/dryad.108679/README.txt?sequence=2"), "raw"),"lobster_strata_2015.tsv")

# with built-in defaults:
lobster <- whoa::read_whoa(data = "lobster_data_2015.vcf")

# Using more arguments: keeping only samples in BUZ sampling sites and using 2000 SNPs:
indivs_buz <- readr::read_tsv(file = "lobster_strata_2015.tsv",
                          col_names = c("INDIVIDUALS", "STRATA"),
                          col_types = "cc") \%>\%
  dplyr::filter(STRATA == "7") \%>\%
  dplyr::select(INDIVIDUALS) \%>\%
  purrr::flatten_chr(.)

lobster_buz_2000 <- whoa::read_whoa(
    data = "lobster_data_2015.vcf",
    indivs = indivs_buz,
    sample.snps = 2000,
    random.seed = 62994,
    filename = "lobster_buz_2000")
}
}
\references{
Zheng X, Gogarten S, Lawrence M, Stilp A, Conomos M, Weir BS,
Laurie C, Levine D (2017). SeqArray -- A storage-efficient high-performance
data format for WGS variant calls.
Bioinformatics.

Danecek P, Auton A, Abecasis G et al. (2011)
The variant call format and VCFtools.
Bioinformatics, 27, 2156-2158.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
