% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoa.R
\docType{package}
\name{whoa}
\alias{whoa}
\alias{whoa-package}
\title{whoa: Evaluation of genotyping error in genotype by sequencing data}
\description{
The name is found in the capitals here:  \code{W}here's my \code{H}eterozygotes?! \code{O}bservations
of genotyping \code{A}ccuracy.
}
\section{the \code{whoa} main high-level functions}{


Fill in.
}

\section{genetic data formats}{


Fill in.
}

\section{example data}{


Fill in.
}

